+++ App/app/src/main/java/com/example/geoagenda/MainActivity.kt
@@ -4,7 +4,9 @@ import android.content.Context
 import android.content.Intent
 import android.content.SharedPreferences
 import android.os.Bundle
+import android.util.Log
 import android.view.Menu
+import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import androidx.drawerlayout.widget.DrawerLayout
@@ -42,7 +44,7 @@ class MainActivity : AppCompatActivity() {
         val prefs =  getSharedPreferences(getString(R.string.prefs_file), Context.MODE_PRIVATE)
         prefs.edit().putString("email", email).commit()
         prefs.edit().putString("provider", provider).commit()
-
+        Toast.makeText(this@MainActivity, prefs.getString("email", "null"), Toast.LENGTH_SHORT).show()
         val drawerLayout: DrawerLayout = findViewById(R.id.drawer_layout)
         val navView: NavigationView = findViewById(R.id.nav_view)
         val navController = findNavController(R.id.nav_host_fragment)
