+++ App/app/src/main/res/layout/activity_create_account.xml
@@ -115,27 +115,13 @@
         app:layout_constraintTop_toBottomOf="@+id/repeatNewPasswordEditTextLayout"
         app:layout_constraintVertical_bias="0.15" />
 
-    <TextView
-        android:id="@+id/createAccountOr"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="201dp"
-        android:layout_marginTop="19dp"
-        android:layout_marginEnd="201dp"
-        android:layout_marginBottom="16dp"
-        android:text="@string/create_account_o"
-        app:layout_constraintBottom_toTopOf="@+id/googleCreateButton"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/createAccountButton" />
-
     <Button
-        android:id="@+id/googleCreateButton"
+        android:id="@+id/discardButton"
         style="?attr/materialButtonStyle"
         android:layout_width="159dp"
         android:layout_height="55dp"
         android:backgroundTint="@color/colorPrimaryDark"
-        android:text="@string/google_create_button"
+        android:text="@string/create_account_cancel"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="@+id/repeatNewPasswordEditTextLayout"
         app:layout_constraintHorizontal_bias="1.0"
