+++ App/app/src/main/res/layout/activity_login.xml
@@ -118,4 +118,18 @@
         app:layout_constraintTop_toBottomOf="@+id/passwordEditTextLayout"
         app:layout_constraintVertical_bias="0.229" />
 
+    <Button
+        android:id="@+id/googleCreateButton"
+        style="?attr/materialButtonStyle"
+        android:layout_width="333dp"
+        android:layout_height="55dp"
+        android:backgroundTint="@color/colorPrimaryDark"
+        android:text="@string/google_create_button"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="@+id/signup_button"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toStartOf="@+id/login_button"
+        app:layout_constraintTop_toBottomOf="@+id/login_button"
+        app:layout_constraintVertical_bias="0.277" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
