+++ App/app/src/main/res/values/strings.xml
@@ -68,8 +68,8 @@
     <string name="create_password_text_field">Contraseña</string>
     <string name="repeat_password_text_field">Repetir contraseña</string>
     <string name="register_button">Crear cuenta</string>
-    <string name="google_create_button">Usar cuenta Google</string>
-    <string name="create_account_o">O</string>
+    <string name="google_create_button">Utilizar cuenta de Google</string>
+    <string name="create_account_cancel">Cancelar</string>
 
     //Para crear lugar
     <string name="create_location">Agregar lugar</string>
